fix(Local Autopilot & Loop Engine): money columns are BIGINT micro-cents; runner stage clocks; builder asks
- Contract bug caught by WP2: costs are MICRO-CENTS (round(usd*100*1e6) = 1e8 per USD), so the
  "$100" DEFAULT_SPEND_CEILING_MICROS was $1, and a true $100 (1e10) overflows a 32-bit Int (~$21.47)
  - an overflowing write fails the whole drive row and saveDriveRow swallows it. LoopDrive.
  spendCeilingMicros and LoopDirection.budgetMicros/usedMicros are BIGINT; the wire stays numbers
  (Number()/BigInt() at the store boundary); default = $100; the storable-max guard becomes a
  $1,000,000 sanity bound.
- loop-lane.ts: the pristine-tree check is stage `baseline` (not `verifying`) and the commit is stage
  `committing`, so a passive screen tells them apart and stageAt times each (WP4 ask); the lessons
  log line says when the runner KEPT lessons (WP5 ask).
- loop-engine.ts: a retried runner lane is cut from ascent/runner, not HEAD (WP2 ask) - otherwise its
  delivery is refused as a non-fast-forward.

+++ prisma/init.sql
@@ -1592,7 +1592,7 @@ CREATE TABLE "LoopDrive" (
     "mode" TEXT NOT NULL DEFAULT 'bounded',
     "pausedReason" TEXT,
     "pausedUntil" TIMESTAMP(3),
-    "spendCeilingMicros" INTEGER,
+    "spendCeilingMicros" BIGINT,
     "repoStateJson" TEXT NOT NULL DEFAULT '[]',
     "dialsJson" TEXT,
     "lastBeatAt" TIMESTAMP(3),
@@ -1614,7 +1614,7 @@ ALTER TABLE "LoopDrive" ADD COLUMN IF NOT EXISTS "delivery" TEXT;
 ALTER TABLE "LoopDrive" ADD COLUMN IF NOT EXISTS "mode" TEXT NOT NULL DEFAULT 'bounded';
 ALTER TABLE "LoopDrive" ADD COLUMN IF NOT EXISTS "pausedReason" TEXT;
 ALTER TABLE "LoopDrive" ADD COLUMN IF NOT EXISTS "pausedUntil" TIMESTAMP(3);
-ALTER TABLE "LoopDrive" ADD COLUMN IF NOT EXISTS "spendCeilingMicros" INTEGER;
+ALTER TABLE "LoopDrive" ADD COLUMN IF NOT EXISTS "spendCeilingMicros" BIGINT;
 ALTER TABLE "LoopDrive" ADD COLUMN IF NOT EXISTS "repoStateJson" TEXT NOT NULL DEFAULT '[]';
 ALTER TABLE "LoopDrive" ADD COLUMN IF NOT EXISTS "dialsJson" TEXT;
 ALTER TABLE "LoopDrive" ADD COLUMN IF NOT EXISTS "lastBeatAt" TIMESTAMP(3);
@@ -1667,9 +1667,9 @@ CREATE TABLE "LoopDirection" (
     "fenceJson" TEXT NOT NULL DEFAULT '[]',
     "checkText" TEXT NOT NULL DEFAULT '',
     "budgetCycles" INTEGER NOT NULL DEFAULT 3,
-    "budgetMicros" INTEGER,
+    "budgetMicros" BIGINT,
     "usedCycles" INTEGER NOT NULL DEFAULT 0,
-    "usedMicros" INTEGER NOT NULL DEFAULT 0,
+    "usedMicros" BIGINT NOT NULL DEFAULT 0,
     "status" TEXT NOT NULL DEFAULT 'active',
     "originPlanId" TEXT NOT NULL,
     "approvedBy" TEXT,
